requirements.txt
pyicu
numpy
Flask
polyglot
morfessor
langid
newspaper3k
pycld2
gensim
spacy
#summa
#pattern